requirements.txt
aiohttp==3.9.3
aiomultiprocess==0.9.0
aioresponses==0.7.6
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.5.0
EditorConfig==0.12.4
filelock==3.13.1
frozenlist==1.4.1
future==0.18.3
googlesearch-python==1.2.3
httpretty==1.1.4
idna==3.6
jsbeautifier==1.15.0
lxml==5.1.0
multidict==6.0.5
prettytable==3.9.0
python-Wappalyzer==0.3.1
python-whois==0.8.0
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
six==1.16.0
soupsieve==2.5
tldextract==5.1.1
ujson==5.9.0
urllib3==2.2.0
uvloop==0.19.0
wcwidth==0.2.13
wget==3.2
yarl==1.9.4
